requirements.txt
Tk
configparser
customtkinter
CTkListbox
pillow
beautifulsoup4